+++ src/main/scala/stellar/sdk/KeyPair.scala
@@ -176,10 +176,13 @@ object KeyPair extends Decode {
   def fromAddress(address: String)(implicit ec: ExecutionContext): Future[PublicKey] = {
     address match {
       case AddressRegex(name, domain) =>
-        DomainInfo.forDomain(s"https://$domain")
+        DomainInfo.forDomain(s"https://$domain").recoverWith { case t => DomainInfo.forDomain(s"http://$domain") }
           .flatMap {
-            case Some(info) => info.federationServer.byName(s"$name*$domain").map(_.map(_.account))
+            case Some(info) => info.federationServer match {
+              case Some(fedServer) => fedServer.byName(s"$name*$domain").map(_.map(_.account))
                 .map(_.getOrElse(throw NoSuchAddress(address, new Exception(s"Address not found on ${info.federationServer}"))))
+              case _ => Future(throw NoSuchAddress(address, new Exception("Domain info did not contain federation server")))
+            }
             case None => Future(throw NoSuchAddress(address, new Exception("Domain info could not be found")))
           }
           .recover {